requirements.txt
-i https://pypi.org/simple
bertopic==0.13.0
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.0.1
click==8.1.3 ; python_version >= '3.7'
cython==0.29.33 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
filelock==3.9.0 ; python_version >= '3.7'
flask==2.2.2
hdbscan==0.8.29
huggingface-hub==0.11.1 ; python_full_version >= '3.7.0'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version < '3.10'
itsdangerous==2.1.2 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
joblib==1.2.0 ; python_version >= '3.7'
llvmlite==0.39.1 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
nltk==3.8.1 ; python_version >= '3.7'
numba==0.56.4 ; python_version >= '3.7'
numpy==1.23.5 ; python_version >= '3.8'
packaging==23.0 ; python_version >= '3.7'
pandas==1.5.2
pillow==9.4.0 ; python_version >= '3.7'
plotly==5.12.0 ; python_version >= '3.6'
praw==7.6.1
prawcore==2.3.0 ; python_version ~= '3.6'
pynndescent==0.5.8
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==0.21.0
pytz==2022.7.1
pyyaml==5.4.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
regex==2022.10.31 ; python_version >= '3.6'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
scikit-learn==1.2.0 ; python_version >= '3.8'
scipy==1.10.0 ; python_version < '3.12' and python_version >= '3.8'
sentence-transformers==2.2.2 ; python_full_version >= '3.6.0'
sentencepiece==0.1.97
setuptools==66.0.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
tenacity==8.1.0 ; python_version >= '3.6'
threadpoolctl==3.1.0 ; python_version >= '3.6'
tokenizers==0.13.2
torch==1.13.1 ; python_full_version >= '3.7.0'
torchvision==0.14.1 ; python_version >= '3.7'
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==4.25.1 ; python_full_version >= '3.7.0'
typing-extensions==4.4.0 ; python_version >= '3.7'
umap-learn==0.5.3
update-checker==0.18.0
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
websocket-client==1.4.2 ; python_version >= '3.7'
werkzeug==2.2.2 ; python_version >= '3.7'
wheel==0.38.4 ; python_version >= '3.7'
xlsxwriter==3.0.7 ; python_version >= '3.7'
zipp==3.11.0 ; python_version >= '3.7'